Migrate rrq's internal database and storage

Migrate the rrq internal format after breaking changes. We take a "best-effort" approach to migrate data, but generally one should not assume that the rrq data store is a persistent store. If you try to connect to a database that has not been migrated, an error will be thrown indicating how you can run this function.

rrq_migrate(queue_id, con = redux::hiredis())

Arguments

queue_id

The name of the rrq queue (as passed to rrq_controller

con

A redis connection

Value

Nothing, called for its side effects

Compatibility

We can detect if a rrq database has been opened with a newer version of rrq and attempt to fix it. However, older versions do not now know when the database has become out of date. Using the same version of worker and controller will generally give good results though.

0.3.1 to 0.4.0

We moved from an internal data store based on storr to one that allows offloading. This requires moving all data around to the new store and rewriting tasks to reference it (both task inputs and task outputs).