Wait for group of tasks

Poll for a group of tasks to complete, returning the result as list when completed. If the tasks have already completed this is roughly equivalent to tasks_result.

rrq_tasks_wait(
  task_ids,
  timeout = NULL,
  time_poll = 1,
  progress = NULL,
  error = FALSE,
  follow = NULL,
  controller = NULL
)

Arguments

task_ids

A vector of task ids to poll for

timeout

Optional timeout, in seconds, after which an error will be thrown if the task has not completed. If not given, falls back on the controller's timeout_task_wait (see rrq_controller2)

time_poll

Optional time with which to "poll" for completion (default is 1s, see rrq_task_wait() for details)

progress

Optional logical indicating if a progress bar should be displayed. If NULL we fall back on the value of the global option rrq.progress, and if that is unset display a progress bar if in an interactive session.

error

Logical, indicating if we should throw an error if the task was not successful. See rrq_task_result() for details. Note that an error is always thrown if not all tasks are fetched in time.

follow

Optional logical, indicating if we should follow any redirects set up by doing rrq_task_retry. If not given, falls back on the value passed into the controller, the global option rrq.follow, and finally TRUE. Set to FALSE if you want to return information about the original task, even if it has been subsequently retried.

controller

The controller to use. If not given (or NULL) we'll use the controller registered with rrq_default_controller_set().

Value

A list with task values, but we will change this soon