rrq queue controller

A queue controller. Use this to interact with a queue/cluster.

Task lifecycle

A task is queued with $enqueue(), at which point it becomes PENDING

Once a worker selects the task to run, it becomes RUNNING

If the task completes successfully without error it becomes COMPLETE

If the task throws an error, it becomes ERROR

If the task was cancelled (e.g., via $task_cancel()) it becomes CANCELLED

If the task is killed by an external process, crashes or the worker dies (and is running a heartbeat) then the task becomes DIED.

The status of an unknown task is MISSING

Tasks in any terminal state (except IMPOSSIBLE) may be retried with task_retry at which point they become MOVED, see vignette("fault-tolerance") for details

Worker lifecycle

A worker appears and is IDLE

When running a task it is BUSY

If it receives a PAUSE message it becomes PAUSED until it receives a RESUME message

If it exits cleanly (e.g., via a STOP message or a timeout) it becomes EXITED

If it crashes and was running a heartbeat, it becomes LOST

Messages

Most of the time workers process tasks, but you can also send them "messages". Messages take priority over tasks, so if a worker becomes idle (by coming online or by finishing a task) it will consume all available messages before starting on a new task, even if both are available.

Each message has a "command" and may have "arguments" to that command. The supported messages are:

PING (no args): "ping" the worker, if alive it will respond with "PONG"

ECHO (accepts an argument of a string): Print a string to the terminal and log of the worker. Will respond with OK once the message has been printed.

EVAL (accepts a string or a quoted expression): Evaluate an arbitrary R expression on the worker. Responds with the value of this expression.

STOP (accepts a string to print as the worker exits, defaults to "BYE"): Tells the worker to stop.

INFO (no args): Returns information about the worker (versions of packages, hostname, pid, etc).

PAUSE (no args): Tells the worker to stop accepting tasks (until it receives a RESUME message). Messages are processed as normal.

RESUME (no args): Tells a paused worker to resume accepting tasks.

REFRESH (no args): Tells the worker to rebuild their environment with the create method.

TIMEOUT_SET (accepts a number, representing seconds): Updates the worker timeout - the length of time after which it will exit if it has not processed a task.

TIMEOUT_GET (no args): Tells the worker to respond with its current timeout.

Public fields

con

The redis connection. This is part of the public API and can be used to access the same redis database as the queue.

queue_id

The queue id used on creation. This is read-only after creation.

Methods

Public methods

rrq_controller$new()

rrq_controller$to_v2()

Method new()

Constructor

Usage

rrq_controller$new(
  queue_id,
  con = redux::hiredis(),
  timeout_task_wait = NULL,
  follow = NULL,
  check_version = TRUE
)

Arguments

queue_id

An identifier for the queue. This will prefix all keys in redis, so a prefix might be useful here depending on your use case (e.g. rrq:<user>:<id>)

con

A redis connection. The default tries to create a redis connection using default ports, or environment variables set as in redux::hiredis()

timeout_task_wait

An optional default timeout to use when waiting for tasks (e.g., with $task_wait(), $tasks_wait(), etc). If not given, then we fall back on the global option rrq.timeout_task_wait, and if that is not set,

follow

An optional default logical to use for tasks that may (or may not) be retried. If not given we fall back on the global option rrq.follow, and if that is not set then TRUE (i.e., we do follow). The value follow = TRUE is potentially slower than follow = FALSE for some operations because we need to dereference every task id. If you never use $task_retry then this dereference never has an effect and we can skip it. See vignette("fault-tolerance") for more information.

check_version

Check that the schema version is correct

Method to_v2()

Convert controller to the new-style object. Please don't use this in packages directly

Usage

rrq_controller$to_v2()