rrq queue worker

rrq queue worker

rrq queue worker

Details

A rrq queue worker. These are not typically for interacting with but will sit and poll a queue for jobs.

Public fields

name

The name of the worker

Methods

Public methods

rrq_worker$new()

rrq_worker$info()

rrq_worker$log()

rrq_worker$load_envir()

rrq_worker$poll()

rrq_worker$step()

rrq_worker$loop()

rrq_worker$format()

rrq_worker$timer_start()

rrq_worker$progress()

rrq_worker$task_eval()

rrq_worker$shutdown()

Method new()

Constructor

Usage

rrq_worker$new(
  queue_id,
  con = redux::hiredis(),
  key_alive = NULL,
  worker_name = NULL,
  queue = NULL,
  time_poll = NULL,
  timeout = NULL,
  heartbeat_period = NULL,
  verbose = TRUE,
  register = TRUE
)

Arguments

queue_id

Name of the queue to connect to. This will be the prefix to all the keys in the redis database

con

A redis connection

key_alive

Optional key that will be written once the worker is alive.

worker_name

Optional worker name. If omitted, a random name will be created.

queue

Queues to listen on, listed in decreasing order of priority. If not given, we listen on the "default" queue.

time_poll

Poll time. Longer values here will reduce the impact on the database but make workers less responsive to being killed with an interrupt (control-C or Escape). The default should be good for most uses, but shorter values are used for debugging.

timeout

Optional timeout to set for the worker. This is (roughly) equivalent to issuing a TIMEOUT_SET message after initialising the worker, except that it's guaranteed to be run by all workers.

heartbeat_period

Optional period for the heartbeat. If non-NULL then a heartbeat process will be started (using heartbeat) which can be used to build fault tolerant queues.

verbose

Logical, indicating if the worker should print logging output to the screen. Logging to screen has a small but measurable performance cost, and if you will not collect system logs from the worker then it is wasted time. Logging to the redis server is always enabled.

register

Logical, indicating if the worker should be registered with the controller. Typically this is TRUE, but set to FALSE if you need to impersonate a worker, or create a temporary worker. This is passed as FALSE when running a task in a separate process.

Method info()

Return information about this worker, a list of key-value pairs.

Usage

rrq_worker$info()

Method log()

Create a log entry. This will print a human readable format to screen and a machine-readable format to the redis database.

Usage

rrq_worker$log(label, value = NULL)

Arguments

label

Scalar character, the title of the log entry

value

Character vector (or null) with log values

Method load_envir()

Load the worker environment by creating a new environment object and running the create hook (if configured). See $envir on rrq_controller for details.

Usage

rrq_worker$load_envir()

Method poll()

Poll for work

Usage

rrq_worker$poll(immediate = FALSE)

Arguments

immediate

Logical, indicating if we should not do a blocking wait on the queue but instead reducing the timeout to zero. Intended primarily for use in the tests.

Method step()

Take a single "step". This consists of

Poll for work ($poll())

If work found, run it (either a task or a message)

If work not found, check the timeout

Usage

rrq_worker$step(immediate = FALSE)

Arguments

immediate

Logical, indicating if we should not do a blocking wait on the queue but instead reducing the timeout to zero. Intended primarily for use in the tests.

Method loop()

The main worker loop. Use this to set up the main worker event loop, which will continue until exiting (via a timeout or message).

Usage

rrq_worker$loop(immediate = FALSE)

Arguments

immediate

Logical, indicating if we should not do a blocking wait on the queue but instead reducing the timeout to zero. Intended primarily for use in the tests.

Method format()

Create a nice string representation of the worker. Used automatically to print the worker by R6.

Usage

rrq_worker$format()

Method timer_start()

Start the timer

Usage

rrq_worker$timer_start()

Method progress()

Submit a progress message. See rrq_task_progress_update() for details of this mechanism.

Usage

rrq_worker$progress(value, error = TRUE)

Arguments

value

An R object with the contents of the update. This will overwrite any previous progress value, and can be retrieved from a rrq_controller with the $task_progress method. A value of NULL will appear to clear the status, as NULL will also be returned if no status is found for a task.

error

Logical, indicating if we should throw an error if not running as an rrq task. Set this to FALSE if you want code to work without modification within and outside of an rrq job, or to TRUE if you want to be sure that progress messages have made it to the server.

Method task_eval()

Evaluate a task

Usage

rrq_worker$task_eval(task_id)

Arguments

task_id

A task identifier. It is undefined what happens if this identifier does not exist.

Method shutdown()

Stop the worker

Usage

rrq_worker$shutdown(status = "OK", graceful = TRUE)

Arguments

status

the worker status; typically be one of OK or ERROR but can be any string

graceful

Logical, indicating if we should request a graceful shutdown of the heartbeat, if running.